# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing villagers, donations, loans, and repayments.",
        "title": "UEM Syariah API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/villagers": {
            "get": {
                "description": "Retrieve a list of all villagers",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villagers"
                ],
                "summary": "Get all villagers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Villager"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Add a villager with name, family head, and neighborhood",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villagers"
                ],
                "summary": "Create a new villager",
                "parameters": [
                    {
                        "description": "Villager Data",
                        "name": "villager",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Villager"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Villager"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/villagers/{id}": {
            "get": {
                "description": "Retrieve a single villager by their ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villagers"
                ],
                "summary": "Get a villager by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Villager ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Villager"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "put": {
                "description": "Update villager details by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villagers"
                ],
                "summary": "Update a villager",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Villager ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated Villager Data",
                        "name": "villager",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Villager"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Villager"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Remove a villager by ID",
                "tags": [
                    "villagers"
                ],
                "summary": "Delete a villager",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Villager ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/villagers/{id}/loans": {
            "get": {
                "description": "Retrieve all loans associated with a villager by their ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "villagers"
                ],
                "summary": "Get loans for a specific villager",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Villager ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Loan"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Loan": {
            "type": "object",
            "required": [
                "amount"
            ],
            "properties": {
                "actual_end_date": {
                    "type": "string"
                },
                "amount": {
                    "type": "number"
                },
                "created_at": {
                    "type": "string"
                },
                "current_amount_paid": {
                    "type": "number"
                },
                "deleted_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "planned_end_date": {
                    "type": "string"
                },
                "reason": {
                    "description": "optional",
                    "type": "string"
                },
                "rest_payment": {
                    "type": "number"
                },
                "start_date": {
                    "type": "string"
                },
                "status": {
                    "description": "\"ongoing\", \"paid\", \"defaulted\"",
                    "type": "string"
                },
                "total_amount_paid": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                },
                "villager_id": {
                    "description": "borrower",
                    "type": "integer"
                }
            }
        },
...
        "models.Villager": {
            "type": "object"
        }
    }
}